requirements.txt
django                      # The Django web framework
djangorestframework         # For building RESTful APIs
djangorestframework-simplejwt   # For JWT authentication
simplejson          # For enhanced JSON handling
django-cors-headers # For handling CORS in Django